webapi/webapi-iap-xwalk-tests/iap/Navigator_iap_purchase.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WebAPI: IAP</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Zhong Qiu" href="mailto:zhongx.qiu@intel.com">
<link rel="help" href="http://www.w3.org/TR/IAP-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  at = async_test("Check if purchase method works well");
  navigator.iap.init(options).then(function() {

    at.step(function() {
      var appleProductsId = "org.xwalk.orange";
      var queryProducts = {"productId": appleProductsId,
                          count: 1,
                          "userInfo": "Buy an orange!"};
      navigator.iap.purchase(queryProducts).then(function(result) {
        assert_true(result instanceof Object, "The returned products info should be an array");

        test(function() {
          assert_true("productId" in result, "PurchaseResult object has member productId");
          assert_own_property(result, "productId", "PurchaseResult object has own property productId");
          assert_equals(typeof result.productId, "string", "productId property should be a string");
        }, "Check the productId property");
        test(function() {
          assert_true("transactionId" in result, "PurchaseResult object has member transactionId");
          assert_own_property(result, "transactionId", "PurchaseResult object has own property description");
          assert_equals(typeof result.transactionId, "string", "transactionId property should be a string");
        }, "Check the transactionId property");
        test(function() {
          assert_true("productReceipt" in result, "PurchaseResult object has member productReceipt");
          assert_own_property(result, "productReceipt", "PurchaseResult object has own property productReceipt");
          assert_equals(typeof result.productReceipt, "string", "productReceipt property should be a string");
        }, "Check the productReceipt property");
        test(function() {
          assert_true("packageName" in result, "PurchaseResult object has member packageName");
          assert_own_property(result, "packageName", "PurchaseResult object has own property packageName");
          assert_equals(typeof result.packageName, "string", "packageName property should be a string");
        }, "Check the packageName property");
        test(function() {
          assert_true("purchaseTime" in result, "PurchaseResult object has member purchaseTime");
          assert_own_property(result, "purchaseTime", "PurchaseResult object has own property purchaseTime");
          assert_equals(typeof result.purchaseTime, "string", "purchaseTime property should be a string");
        }, "Check the purchaseTime property");
        test(function() {
          assert_true("purchaseState" in result, "PurchaseResult object has member purchaseState");
          assert_own_property(result, "purchaseState", "PurchaseResult object has own property purchaseState");
          assert_equals(typeof result.purchaseState, "string", "purchaseState property should be a string");
        }, "Check the purchaseState property");

        at.done();
      }).catch(function(e) {
        assert_unreached("Should not be here, otherwise purchase failed!");
        at.done();
      });
    });
  }).catch(function(e) {
    assert_unreached("Should not be here, otherwise init failed!");
    at.done();
  });
</script>
